requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.10
django-cors-headers==3.10.1
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
hubspot-api-client==4.0.6
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
